Updting README and Cmake file to reflect names changes
