--- Code_Exercises/Image_Convolution/README.md
@@ -1,21 +1,0 @@
-# SYCL Academy
-
-## Image Convolution
----
-
-This code is used as a walkthrough during ISC24 to help you understand a more 
-complex way to write SYCL code using Functors and other techniques like 
-nd-range.
-
-This is based on the exercises in the main branch of the SYCL Academy project 
-so you can re-visit this after ISC to learn more about SYCL development.
-
----
-
-## Build and execution hints
-
-For DevCloud follow these [instructions](../devcloud.md).
-
-For DPC++: [instructions](../dpcpp.md).
-
-For AdaptiveCpp: [instructions](../adaptivecpp.md).
