+++ Code_Exercises/Image_Convolution_Functors/source.cpp
@@ -1,0 +1,17 @@
+/*
+ SYCL Academy (c)
+
+ SYCL Academy is licensed under a Creative Commons
+ Attribution-ShareAlike 4.0 International License.
+
+ You should have received a copy of the license along with this
+ work.  If not, see <http://creativecommons.org/licenses/by-sa/4.0/>.
+*/
+
+#define CATCH_CONFIG_MAIN
+#include <catch2/catch.hpp>
+
+TEST_CASE("image_convolution_1D", "1_source") {
+  // Use your code from Exercise Vectors to start
+  REQUIRE(true);
+}
