+++ Lesson_Materials/Introduction_to_SYCL/index.html
@@ -585,7 +585,7 @@ handler.parallel_for(range, func);
 					</div>
 					<div class="container" data-markdown>
             * Use Data Parallelism and write a SYCL kernel
-            * The exercise README is in the "Code_Exercises/Data_Parallelism" folder
+            * The exercise README is in the "Code_Exercises/Introduction_to_SYCL" folder
             * Follow the guidelines in the README and comments in the source.cpp file
 	    * There is a solution file but only use it if you need to
 					</div>
@@ -638,7 +638,7 @@ handler.parallel_for(range, func);
 					</div>
 					<div class="container" data-markdown>
             * Explain a more complex example through image convolution
-            * Solution available in `Code_Exercises/Image_Convolution/solution.cpp`
+            * Solution available in `Code_Exercises/Image_Convolution_Functors/solution.cpp`
 					</div>
 				</section>
 				<!--Slide 39-->
