+++ README.md
@@ -45,7 +45,7 @@ below to find the necessary files. At the command line just remember that the
 
 | Lesson | Title | Slides | Exercise | Source | Solution | DPC++ | AdaptiveCpp |
 |--------|-------|--------|----------|--------|----------|-------|---------|
-| 01 | What is SYCL | [slides][lesson-1-slides] | [exercise][lesson-1-exercise] | [source][lesson-1-source] | [solution][lesson-1-solution] | Yes | Yes |
+| 01 | Introduction to SYCL | [slides][lesson-1-slides] | [exercise][lesson-1-exercise] | [source][lesson-1-source] | [solution][lesson-1-solution] | Yes | Yes |
 | 02 | ND Range Kernels | [slides][lesson-14-slides] | [exercise][lesson-14-exercise] | [source][lesson-14-source] | [solution][lesson-14-solution] | Yes | Yes |
 | 03 | oneMKL GEMM | [slides][lesson-20-slides] | [exercise][lesson-20-exercise] | [source][lesson-20-source] | [solution][lesson-20-solution] | Yes | Yes |
 | Walkthrough | Image Convolution | NA | [exercise][lesson-22-exercise] | [source][lesson-22-source] | [solution][lesson-22-solution] | Yes | Yes |
@@ -301,6 +301,6 @@ The CMake configuration can also be used to build the exercises, see the section
 [lesson-20-source]:   ./Code_Exercises/OneMKL_gemm/source.cpp
 [lesson-20-solution]: ./Code_Exercises/OneMKL_gemm/solution.cpp
 
-[lesson-22-exercise]: ./Code_Exercises/Functors/README.md
-[lesson-22-source]:   ./Code_Exercises/Functors/source.cpp
-[lesson-22-solution]: ./Code_Exercises/Functors/solution.cpp
+[lesson-22-exercise]: ./Code_Exercises/Image_Convolution_Functors/README.md
+[lesson-22-source]:   ./Code_Exercises/Image_Convolution_Functors/source.cpp
+[lesson-22-solution]: ./Code_Exercises/Image_Convolution_Functors/solution.cpp
